requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pydub>=0.25.1
google-cloud-speech>=2.21.0